+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 * Registration module: check no duplicated output when coaligning files.
 * Registration module: fix selection of files to coalign (e.g. when registering image1_BF.ome.tif, do not coalign image10_BF.ome.tif).
 * Check images/masks are valid before starting processing (to avoid problems with parallel processing).
-
+* Registration module (alignment): check distinct input files generate distinct output files.